fix: add auth api type
#5
